Find an associated block

Given a file and line, discover which block the tag is associated with

Usage

associated_block(file, line)

Arguments

file

(character[1]) A file path associated with a tag

line

(integer[1]) A line number in the file where the tag was found

Value

(roxygen2::roxy_block() | NULL) A roxy_block if one could be associated, or NULL if not.